+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ## [Unreleased]
 
+### Changed
+- 對調所有列表頁 DFP 與 OneAdICIP 的位置 @esbb48
+
 ## 1.0.48 - 2017-12-22
 ### Added
 - 增加新聞頁時對 sn 型態的判斷 @esbb48